+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 - Add test-related sandbox configurations and runtime fixes
 - Set the required `extracted` and `collected` attributes for custom test classpath configurations
+- Set the repository name and fix the `action` application in `createCustomPluginRepository`
 
 ## [2.5.0] - 2025-04-01